Core/BattleGround: Move BattleGroundQueue to own class. Update BattleGroundQueue to TrinityCore.
* Includes simplified Arena team select code
