+++ src/game/BattleGround/BattleGroundQueue.h
@@ -1,0 +1,177 @@
+/*
+ * This file is part of the CMaNGOS Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ */
+
+#ifndef __BATTLEGROUNDQUEUE_H
+#define __BATTLEGROUNDQUEUE_H
+
+#include "Common.h"
+#include "DBCEnums.h"
+#include "BattleGround.h"
+
+#include <mutex>
+
+#define COUNT_OF_PLAYERS_TO_AVERAGE_WAIT_TIME 10
+
+struct GroupQueueInfo;                                      // type predefinition
+struct PlayerQueueInfo                                      // stores information for players in queue
+{
+    uint32  LastOnlineTime;                                 // for tracking and removing offline players from queue after 5 minutes
+    GroupQueueInfo* GroupInfo;                              // pointer to the associated groupqueueinfo
+};
+
+typedef std::map<ObjectGuid, PlayerQueueInfo*> GroupQueueInfoPlayers;
+
+struct GroupQueueInfo                                       // stores information about the group in queue (also used when joined as solo!)
+{
+    GroupQueueInfoPlayers Players;                          // player queue info map
+    Team  GroupTeam;                                        // Player team (ALLIANCE/HORDE)
+    BattleGroundTypeId BgTypeId;                            // battleground type id
+    bool    IsRated;                                        // rated
+    ArenaType arenaType;                                    // 2v2, 3v3, 5v5 or 0 when BG
+    uint32  ArenaTeamId;                                    // team id if rated match
+    uint32  JoinTime;                                       // time when group was added
+    uint32  RemoveInviteTime;                               // time when we will remove invite for players in group
+    uint32  IsInvitedToBGInstanceGUID;                      // was invited to certain BG
+    uint32  ArenaTeamRating;                                // if rated match, inited to the rating of the team
+    uint32  OpponentsTeamRating;                            // for rated arena matches
+};
+
+enum BattleGroundQueueGroupTypes
+{
+    BG_QUEUE_PREMADE_ALLIANCE   = 0,
+    BG_QUEUE_PREMADE_HORDE      = 1,
+    BG_QUEUE_NORMAL_ALLIANCE    = 2,
+    BG_QUEUE_NORMAL_HORDE       = 3
+};
+#define BG_QUEUE_GROUP_TYPES_COUNT 4
+
+class BattleGround;
+class BattleGroundQueue
+{
+    public:
+        BattleGroundQueue();
+        ~BattleGroundQueue();
+
+        void BattleGroundQueueUpdate(BattleGroundTypeId bgTypeId, BattleGroundBracketId bracket_id, ArenaType arenaType = ARENA_TYPE_NONE, bool isRated = false, uint32 minRating = 0);
+
+        void FillPlayersToBG(BattleGround* bg, BattleGroundBracketId bracket_id);
+        bool CheckPremadeMatch(BattleGroundBracketId bracket_id, uint32 MinPlayersPerTeam, uint32 MaxPlayersPerTeam);
+        bool CheckNormalMatch(BattleGround* bg_template, BattleGroundBracketId bracket_id, uint32 minPlayers, uint32 maxPlayers);
+        bool CheckSkirmishForSameFaction(BattleGroundBracketId bracket_id, uint32 minPlayersPerTeam);
+        GroupQueueInfo* AddGroup(Player* leader, Group* group, BattleGroundTypeId bgTypeId, BattleGroundBracketId bracketId, ArenaType arenaType, bool isRated, bool isPremade, uint32 ArenaRating, uint32 ArenaTeamId = 0);
+        void RemovePlayer(ObjectGuid guid, bool decreaseInvitedCount);
+        bool IsPlayerInvited(ObjectGuid pl_guid, const uint32 bgInstanceGuid, const uint32 removeTime);
+        bool GetPlayerGroupInfoData(ObjectGuid guid, GroupQueueInfo* ginfo);
+        void PlayerInvitedToBGUpdateAverageWaitTime(GroupQueueInfo* ginfo, BattleGroundBracketId bracket_id);
+        uint32 GetAverageQueueWaitTime(GroupQueueInfo* ginfo, BattleGroundBracketId bracket_id);
+
+    private:
+        // mutex that should not allow changing private data, nor allowing to update Queue during private data change.
+        std::recursive_mutex m_Lock;
+
+
+        typedef std::map<ObjectGuid, PlayerQueueInfo> QueuedPlayersMap;
+        QueuedPlayersMap m_QueuedPlayers;
+
+        // we need constant add to begin and constant remove / add from the end, therefore deque suits our problem well
+        typedef std::list<GroupQueueInfo*> GroupsQueueType;
+
+        /*
+        This two dimensional array is used to store All queued groups
+        First dimension specifies the bgTypeId
+        Second dimension specifies the player's group types -
+             BG_QUEUE_PREMADE_ALLIANCE  is used for premade alliance groups and alliance rated arena teams
+             BG_QUEUE_PREMADE_HORDE     is used for premade horde groups and horde rated arena teams
+             BG_QUEUE_NORMAL_ALLIANCE   is used for normal (or small) alliance groups or non-rated arena matches
+             BG_QUEUE_NORMAL_HORDE      is used for normal (or small) horde groups or non-rated arena matches
+        */
+        GroupsQueueType m_QueuedGroups[MAX_BATTLEGROUND_BRACKETS][BG_QUEUE_GROUP_TYPES_COUNT];
+
+        // class to select and invite groups to bg
+        class SelectionPool
+        {
+            public:
+                SelectionPool() : PlayerCount(0) { }
+                void Init();
+                bool AddGroup(GroupQueueInfo* ginfo, uint32 desiredCount);
+                bool KickGroup(uint32 size);
+                uint32 GetPlayerCount() const { return PlayerCount; }
+            public:
+                GroupsQueueType SelectedGroups;
+            private:
+                uint32 PlayerCount;
+        };
+
+        // one selection pool for horde, other one for alliance
+        SelectionPool m_SelectionPools[PVP_TEAM_COUNT];
+    private:
+
+        bool InviteGroupToBG(GroupQueueInfo* ginfo, BattleGround* bg, Team side);
+        uint32 m_WaitTimes[PVP_TEAM_COUNT][MAX_BATTLEGROUND_BRACKETS][COUNT_OF_PLAYERS_TO_AVERAGE_WAIT_TIME];
+        uint32 m_WaitTimeLastPlayer[PVP_TEAM_COUNT][MAX_BATTLEGROUND_BRACKETS];
+        uint32 m_SumOfWaitTimes[PVP_TEAM_COUNT][MAX_BATTLEGROUND_BRACKETS];
+};
+
+/*
+    This class is used to invite player to BG again, when minute lasts from his first invitation
+    it is capable to solve all possibilities
+*/
+class BGQueueInviteEvent : public BasicEvent
+{
+    public:
+        BGQueueInviteEvent(ObjectGuid pl_guid, uint32 BgInstanceGUID, BattleGroundTypeId BgTypeId, ArenaType arenaType, uint32 removeTime) :
+            m_PlayerGuid(pl_guid), m_BgInstanceGUID(BgInstanceGUID), m_BgTypeId(BgTypeId), m_ArenaType(arenaType), m_RemoveTime(removeTime)
+        {
+        };
+        virtual ~BGQueueInviteEvent() {};
+
+        virtual bool Execute(uint64 e_time, uint32 p_time) override;
+        virtual void Abort(uint64 e_time) override;
+    private:
+        ObjectGuid m_PlayerGuid;
+        uint32 m_BgInstanceGUID;
+        BattleGroundTypeId m_BgTypeId;
+        ArenaType m_ArenaType;
+        uint32 m_RemoveTime;
+};
+
+/*
+    This class is used to remove player from BG queue after 1 minute 20 seconds from first invitation
+    We must store removeInvite time in case player left queue and joined and is invited again
+    We must store bgQueueTypeId, because battleground can be deleted already, when player entered it
+*/
+class BGQueueRemoveEvent : public BasicEvent
+{
+    public:
+        BGQueueRemoveEvent(ObjectGuid plGuid, uint32 bgInstanceGUID, BattleGroundTypeId BgTypeId, BattleGroundQueueTypeId bgQueueTypeId, uint32 removeTime)
+            : m_PlayerGuid(plGuid), m_BgInstanceGUID(bgInstanceGUID), m_RemoveTime(removeTime), m_BgTypeId(BgTypeId), m_BgQueueTypeId(bgQueueTypeId)
+        {}
+
+        virtual ~BGQueueRemoveEvent() {}
+
+        virtual bool Execute(uint64 e_time, uint32 p_time) override;
+        virtual void Abort(uint64 e_time) override;
+    private:
+        ObjectGuid m_PlayerGuid;
+        uint32 m_BgInstanceGUID;
+        uint32 m_RemoveTime;
+        BattleGroundTypeId m_BgTypeId;
+        BattleGroundQueueTypeId m_BgQueueTypeId;
+};
+
+#endif
